requirements.txt
flask
flask-socketio
scapy
psutil
requests
python-dotenv

# Network analysis
reportlab
openpyxl

# AI Anomaly Detection Engine
scikit-learn
numpy
pandas
joblib
